+++ test/unit/engine/graph/GraphRenderEngineTest.cpp
@@ -500,6 +500,9 @@ namespace GraphRenderEngineTest {
     EXPECT_NE(std::string::npos, message.find("rejectedLeaves=1")) << message;
     EXPECT_NE(std::string::npos, message.find("frustumRejectedLeaves=1")) << message;
     EXPECT_NE(std::string::npos, message.find("backfaceRejectedLeaves=0")) << message;
+    EXPECT_NE(std::string::npos, message.find("tileGrid=1x1")) << message;
+    EXPECT_NE(std::string::npos, message.find("visibleTileReferences=1")) << message;
+    EXPECT_NE(std::string::npos, message.find("uncertainTileLeaves=0")) << message;
   }
 
   TEST(GraphRenderEngine, RecordsRasterVisibilityFrontToBackOrderingMetrics) {